+++ packages/@expo/prebuild-config/src/plugins/withDefaultPlugins.ts
@@ -122,6 +122,8 @@ export const withAndroidExpoPlugins: ConfigPlugin<{
     // If we renamed the package, we should also move it around and rename it in source files
     // Added last to ensure this plugin runs first. Out of tree solutions will mistakenly resolve the package incorrectly otherwise.
     AndroidConfig.Package.withPackageRefactor,
+    // Add the configuration for react-native-edge-to-edge
+    AndroidConfig.EdgeToEdge.withEdgeToEdge,
   ]);
 };